% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/should_be_verified.R
\name{should_be_verified}
\alias{should_be_verified}
\title{Should be verified}
\usage{
should_be_verified(user, token = NULL)
}
\arguments{
\item{user}{Screen name, user ID, or data frame of user data returned by
\url{https://rtweet.info} function, e.g.,
\code{\link[rtweet]{lookup_users}},
\code{\link[rtweet]{get_timeline}},
\code{\link[rtweet]{search_tweets}}.}

\item{token}{Token object to be used in any requests sent to Twitter's REST
API (it's passed to internal rtweet function). API requests are only made
if the input object is a screen name or user ID. The default, \code{NULL},
will look for a token saved as an environment token or will leverage the
rtweet app to request a token on behalf of the user–this requires the user
to be in an interactive session and to have a valid user name and password
with which to log into their Twitter account.}
}
\value{
A named (screen name) numeric vector represented the predicted
  probability of an account being verified
}
\description{
Generates prediction of whether Twitter user(s) should be verified. See
\strong{details} section below for more information on the prediction model
}
\details{
Predictions generated from a boosted logistic regression model
  trained on 24,000 verified and non-verified Twitter accounts. Performance
  on test data was 96% accurate.
}
